LPS-61082 Refactor names
